+++ src/intercom.module.ts
@@ -1,0 +1,45 @@
+import { ModuleWithProviders, NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { IntercomHideDirective } from './directives/hide.directive';
+import { IntercomShowMessagesDirective } from './directives/show-messages.directive';
+import { IntercomShowNewMessageDirective } from './directives/show-new-message.directive';
+import { IntercomShowDirective } from './directives/show.directive';
+import { IntercomShutdownDirective } from './directives/shutdown.directive';
+import { IntercomTrackEventDirective } from './directives/track-event.directive';
+import { Intercom } from './intercom/intercom';
+import { IntercomConfig } from './shared/intercom-config';
+import { loadIntercom } from './util/load-intercom';
+
+
+@NgModule({
+    imports: [
+        RouterModule
+    ],
+    declarations: [
+        IntercomHideDirective,
+        IntercomShowMessagesDirective,
+        IntercomShowNewMessageDirective,
+        IntercomShowDirective,
+        IntercomShutdownDirective,
+        IntercomTrackEventDirective
+    ],
+    exports: [
+        IntercomHideDirective,
+        IntercomShowMessagesDirective,
+        IntercomShowNewMessageDirective,
+        IntercomShowDirective,
+        IntercomShutdownDirective,
+        IntercomTrackEventDirective
+    ]
+})
+export class IntercomModule {
+    static forRoot(config: IntercomConfig): ModuleWithProviders {
+        return {
+            ngModule: IntercomModule,
+            providers: [
+                { provide: IntercomConfig, useValue: config },
+                Intercom
+            ]
+        }
+    }
+}